+++ projects/composablekernel/include/ck/tensor_operation/gpu/grid/gridwise_gemm_multiple_d_wmma_cshuffle.hpp
@@ -62,11 +62,11 @@ __global__ void
     const index_t g_idx = __builtin_amdgcn_readfirstlane(get_block_1d_id() / num_blocks_per_batch);
 
     const auto a_batch_offset =
-        amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_batch.GetAPtrOffset(g_idx)));
+        amd_wave_read_first_lane(compute_ptr_offset_of_batch.GetAPtrOffset(g_idx));
     const auto b_batch_offset =
-        amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_batch.GetBPtrOffset(g_idx)));
+        amd_wave_read_first_lane(compute_ptr_offset_of_batch.GetBPtrOffset(g_idx));
     const auto e_batch_offset =
-        amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_batch.GetEPtrOffset(g_idx)));
+        amd_wave_read_first_lane(compute_ptr_offset_of_batch.GetEPtrOffset(g_idx));
 
     const auto ds_batch_offset = compute_ptr_offset_of_batch.GetDsPtrOffset(g_idx);
 
@@ -158,11 +158,11 @@ __global__ void
     const index_t g_idx = __builtin_amdgcn_readfirstlane(get_block_1d_id() / num_blocks_per_batch);
 
     const long_index_t a_batch_offset =
-        amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_batch.GetAPtrOffset(g_idx)));
+        amd_wave_read_first_lane(compute_ptr_offset_of_batch.GetAPtrOffset(g_idx));
     const long_index_t b_batch_offset =
-        amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_batch.GetBPtrOffset(g_idx)));
+        amd_wave_read_first_lane(compute_ptr_offset_of_batch.GetBPtrOffset(g_idx));
     const long_index_t e_batch_offset =
-        amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_batch.GetEPtrOffset(g_idx)));
+        amd_wave_read_first_lane(compute_ptr_offset_of_batch.GetEPtrOffset(g_idx));
 
     const auto ds_batch_offset = compute_ptr_offset_of_batch.GetDsPtrOffset(g_idx);
